+++ demo/draw_demo.py
@@ -18,6 +18,6 @@ def draw():
 
 rb.Game.draw = draw
 
-rb.Time.recurrent_call(1000, lambda: print(rb.Draw._cache_size()))
+rb.Time.recurrent_call(lambda: print(rb.Draw._cache_size()), 1000)
 
 rb.begin()