+++ CHANGELOG.md
@@ -19,17 +19,19 @@
 ### Breaking Changes
 
 -   Renamed `ScheduledTask` to `RecurrentTask` and `Time.scheduled_call()` to `Time.recurrent_call()`.
+-   Reordered the arguments of `RecurrentTask`, `FramesTask`, and `DelayedTask` to be more intuitive.
+-   Removed `is_stopped` from the init arguments of above task constructors to simplify them.
 
 ### Added
 
 -   `rigidbody.stop()` method to quickly set the rigidbody velocity to 0.
 -   `Animation` now has an alpha that can be set.
--   Frame, Delayed, and Recurrent tasks now have a `next_run` value that control the next run. This was previously done by `delay`.
 
 ### Changed
 
--   `ScheduledTask` task function can now optionally take in a task argument. This allows you to stop the task when you want.
+-   `RecurrentTask` task callback can now be made to take in a task argument. This allows you to stop the task when you want.
 -   `delay` argument in Frame, Delayed, and Recurrent tasks now only controls the delay.
+-   Frame, Delayed, and Recurrent tasks no longer modify `delay`, and instead determine the next time to run the task using `next_run`.
 
 ### Removed